Use Q8 weights for Klein9B FLUX

+++ src/models/quantized_klein9b/shaders/mul_right_broadcast.py
@@ -33,7 +33,7 @@ MUL_RIGHT_BROADCAST = ShaderVariant(
                 name='y',
                 io_kind=IOKind.INPUT,
                 role='input',
-                contract=TensorContract(dtype='float16', shape=('D',)),
+                contract=TensorContract(dtype='float32', shape=('D',)),
             ),
             TensorFieldSpec(
                 name='output',
@@ -55,17 +55,15 @@ MUL_RIGHT_BROADCAST = ShaderVariant(
     execution_requirements=None,
     source="""\
 #version 450
-#extension GL_EXT_shader_explicit_arithmetic_types_float16 : require
-#extension GL_EXT_shader_16bit_storage : require
 layout(std430) buffer;
 layout(set = 0, binding = 0) buffer restrict readonly XBuffer { float x[]; };
-layout(set = 0, binding = 1) buffer restrict readonly YBuffer { float16_t y[]; };
+layout(set = 0, binding = 1) buffer restrict readonly YBuffer { float y[]; };
 layout(set = 0, binding = 2) buffer restrict writeonly OutputBuffer { float output_values[]; };
 layout(push_constant) uniform PushConstants { uint N; uint H; } pc;
 layout(local_size_x = 256, local_size_y = 1, local_size_z = 1) in;
 void main() {
     const uint idx = gl_GlobalInvocationID.x;
-    if (idx < pc.N) { output_values[idx] = float(x[idx]) * float(y[idx % pc.H]); }
+    if (idx < pc.N) { output_values[idx] = x[idx] * y[idx % pc.H]; }
 }
 """,
 )
